Clean up logging. Extract replaceJsonStringsWithUpdatedIds method
