+++ src/models/document/document-content.ts
@@ -259,11 +259,9 @@ export const DocumentContentModel = DocumentContentModelWithTileDragging.named("
         // Add the model (but not the tile IDs yet) to the Document so that tile references to it won't break
         // when we insert the tiles.
         if (sharedDataSet.id) {
-          console.log("Adding copied sharedModel", sharedDataSet.id, "->", sharedModel.id);
-          console.log("  Which has dataset ids:", sharedDataSet.dataSet?.id, "->", sharedModel.dataSet.id);
           self.addSharedModelFromImport(sharedModel.id, {sharedModel});
         } else {
-          // FIXME can this really happen?
+          // Not sure why this would happen?
           console.warn("Cannot duplicate shared model without an ID");
         }
       }
@@ -283,8 +281,6 @@ export const DocumentContentModel = DocumentContentModelWithTileDragging.named("
 
       // Find the shared models for this tile
       const tileSharedModelEntries = findTileSharedModelEntries(tile.tileId);
-      console.log("tile", tile.tileId, "->", tileIdMap[tile.tileId]);
-      console.log("  has shared models: ", tileSharedModelEntries.map(x => x.sharedModel.id));
 
       // Update the tile's references to its shared models
       const updateFunction = getTileContentInfo(tile.tileType)?.updateContentWithNewSharedModelIds;
@@ -297,7 +293,6 @@ export const DocumentContentModel = DocumentContentModelWithTileDragging.named("
     });
 
     // Add copied tiles to document
-    console.log("calling", insertTileFunction);
     const results = insertTileFunction(updatedTiles, rowInfo);
 
     // Increment default titles when necessary
